# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RequirementCriticality =
  (typeof RequirementCriticality)[keyof typeof RequirementCriticality];

export const RequirementCriticality = {
  must: "must",
  nice: "nice",
} as const;

export type RequirementStatus =
  (typeof RequirementStatus)[keyof typeof RequirementStatus];

export const RequirementStatus = {
  confirmed: "confirmed",
  partial: "partial",
  missing: "missing",
} as const;

export interface Requirement {
  id: number;
  text: string;
  criticality: RequirementCriticality;
  status: RequirementStatus;
}

export interface UserAnswer {
  requirementId: number;
  answer: string;
}

export interface AnalyzeInput {
  vacancyText: string;
  resumeText: string;
}

export interface AnalyzeResult {
  requirements: Requirement[];
}

export interface AdaptInput {
  vacancyText: string;
  resumeText: string;
  userAnswers: UserAnswer[];
}

export interface AdaptResult {
  resumeUpdated: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  ok: boolean;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
}

export interface WaitlistInput {
  email: string;
}

export interface AdminVerifyInput {
  password: string;
}

export interface WaitlistEntry {
  id: number;
  email: string;
  createdAt: string;
}

export interface WaitlistList {
  entries: WaitlistEntry[];
}

export interface UserEntry {
  id: number;
  email: string;
  createdAt: string;
}

export interface UserList {
  users: UserEntry[];
}

export interface CreateUserInput {
  email: string;
  password: string;
}