# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: resume
    description: Resume adaptation operations
  - name: auth
    description: Authentication operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /analyze:
    post:
      operationId: analyzeResume
      tags: [resume]
      summary: Analyze resume against job vacancy
      description: Extracts requirements from job vacancy and matches them against the resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /adapt:
    post:
      operationId: adaptResume
      tags: [resume]
      summary: Adapt resume to job vacancy
      description: Adapts the resume text to match the job vacancy using user answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdaptInput"
      responses:
        "200":
          description: Adapted resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdaptResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/waitlist:
    post:
      operationId: joinWaitlist
      tags: [auth]
      summary: Join the waitlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistInput"
      responses:
        "200":
          description: Added to waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/verify:
    post:
      operationId: verifyAdmin
      tags: [admin]
      summary: Verify admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminVerifyInput"
      responses:
        "200":
          description: Admin verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Wrong password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/waitlist:
    get:
      operationId: getAdminWaitlist
      tags: [admin]
      summary: Get waitlist emails
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Waitlist entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get invited users
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAdminUser
      tags: [admin]
      summary: Create an invited user
      parameters:
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete an invited user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: header
          name: x-admin-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Requirement:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        criticality:
          type: string
          enum: [must, nice]
        status:
          type: string
          enum: [confirmed, partial, missing]
      required:
        - id
        - text
        - criticality
        - status
    UserAnswer:
      type: object
      properties:
        requirementId:
          type: integer
        answer:
          type: string
      required:
        - requirementId
        - answer
    AnalyzeInput:
      type: object
      properties:
        vacancyText:
          type: string
        resumeText:
          type: string
      required:
        - vacancyText
        - resumeText
    AnalyzeResult:
      type: object
      properties:
        requirements:
          type: array
          items:
            $ref: "#/components/schemas/Requirement"
      required:
        - requirements
    AdaptInput:
      type: object
      properties:
        vacancyText:
          type: string
        resumeText:
          type: string
        userAnswers:
          type: array
          items:
            $ref: "#/components/schemas/UserAnswer"
      required:
        - vacancyText
        - resumeText
        - userAnswers
    AdaptResult:
      type: object
      properties:
        resumeUpdated:
          type: string
      required:
        - resumeUpdated
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
      required:
        - id
        - email
    WaitlistInput:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    AdminVerifyInput:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    WaitlistEntry:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - createdAt
    WaitlistList:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/WaitlistEntry"
      required:
        - entries
    UserEntry:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - createdAt
    UserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserEntry"
      required:
        - users
    CreateUserInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password