# lib/api-zod/src/generated/types/userEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserEntry {
  id: number;
  email: string;
  createdAt: string;
}