# lib/api-zod/src/generated/types/waitlistList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WaitlistEntry } from "./waitlistEntry";

export interface WaitlistList {
  entries: WaitlistEntry[];
}